@@ -10,47 +10,43 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.14.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "chai": "^3.5.0",
-    "css-loader": "^0.26.2",
-    "eslint": "^3.9.1",
-    "eslint-plugin-react": "^6.4.1",
-    "extract-text-webpack-plugin": "^2.0.0",
+    "css-loader": "^0.26.4",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^6.10.3",
+    "extract-text-webpack-plugin": "^2.1.2",
+    "file-loader": "^0.11.2",
     "isparta": "^4.0.0",
-    "mocha": "^3.2.0",
-    "sinon": "^1.17.6",
+    "mocha": "^3.4.2",
+    "sinon": "^1.17.7",
     "style-loader": "^0.13.2",
-    "webpack": "^2.2.1"
+    "webpack": "^3.3.0",
+    "webpack-dev-server": "^2.5.1"
   },
   "dependencies": {
-    "body-parser": "^1.15.2",
-    "boom": "^4.1.0",
-    "connect-redis": "^3.1.0",
+    "body-parser": "^1.17.2",
+    "connect-redis": "^3.3.0",
     "cookie-parser": "^1.4.3",
     "csurf": "^1.9.0",
-    "csv": "^1.1.0",
-    "express": "^4.14.0",
-    "express-session": "^1.14.2",
-    "hapi": "^15.1.1",
-    "hapi-auth-cookie": "^6.1.1",
+    "csv": "^1.1.1",
+    "express": "^4.15.3",
+    "express-session": "^1.15.3",
     "immutable": "^3.8.1",
-    "inert": "^4.0.2",
-    "jade": "^1.11.0",
-    "mysql": "^2.12.0",
+    "jsonwebtoken": "^7.4.1",
+    "mysql": "^2.13.0",
     "path": "^0.12.7",
-    "pug": "^2.0.0-beta6",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
+    "pixi.js": "^4.5.5",
+    "pug": "^2.0.0-rc.2",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-micro-container": "^1.1.0",
-    "react-redux": "^5.0.3",
-    "redis": "^2.6.3",
-    "redux": "^3.6.0",
-    "socket.io": "^1.4.8",
-    "uid-safe": "^2.1.3",
-    "vision": "^4.1.0"
+    "redis": "^2.7.1",
+    "socket.io": "^1.7.4",
+    "uid-safe": "^2.1.4"
   }
 }
