@@ -1,62 +1,82 @@
-const unitMaster = require('../data/json/unit.json');
+import { Record, Map } from 'immutable';
+import Controller from './Controller.js';
 
-export default class Client {
+const STATE = Map({
+  LOBBY: 1,
+  PREPARATION: 10,
+  GAME: 20,
+});
 
-  getMetaData(data) {
-    const meta = {};
-    meta.pnum = data.pnum;
-    const deck = {
-      army: [],
-      enemy: []
+export default class Client extends Record({
+  state: STATE.get('LOBBYY'),
+  roomId: null,
+  player: null,
+  opponent: null,
+  controller: null,
+  onUpdate: () => {},
+}) {
+
+  props() {
+    return {
+      roomId: this.roomId,
+      player: this.player,
+      opponent: this.opponent,
+      controller: this.controller,
     };
-    if (data.deck) {
-      Object.keys(data.deck).forEach(pnum => {
-        if (pnum == data.pnum) {
-          deck.army = this.getPrepUnits(data.deck[pnum]);
-        } else {
-          deck.enemy = this.getPrepUnits(data.deck[pnum]);
-        }
-      });
-    }
-    meta.deck = deck;
-    return meta;
   }
 
-  getPrepUnits(ids) {
-    return ids.map(id => {
-      return unitMaster[id];
+  stateIs(str) {
+    return this.state == STATE.get(str);
+  }
+
+  enterRoom({ roomId }) {
+    return this.set('roomId', roomId);
+  }
+
+  leaveRoom() {
+    return this.withMutations(mnt => {
+      mnt.delete('roomId')
+        .delete('controller');
     });
   }
 
-  forecastBattle(game, unit, target, fromCellId, toCellId) {
-    if (!unit || !target || !game.map.isActionable(unit, fromCellId, toCellId)) {
-      return;
-    }
-    const result = {
-      me: {
-        name: unit.klass.name,
-        hp: unit.hp
-      },
-      tg: {
-        name: target.klass.name,
-        hp: target.hp
-      }
-    };
-    if (unit.klass.healer) {
-      result.me.val = unit.klass.pow;
-    } else {
-      result.me.val = unit.effect(target);
-      result.me.hit = unit.hitRate(target, game.map.field.avoid(toCellId));
-      result.me.crit = unit.critRate(target);
-    }
-    if (!unit.klass.healer && !target.klass.healer) {
-      if (game.map.isActionable(target, toCellId, fromCellId)) {
-        result.tg.val = target.effect(unit);
-        result.tg.hit = target.hitRate(unit, game.map.field.avoid(fromCellId));
-        result.tg.crit = target.critRate(unit);
-      }
-    }
-    return result;
+  startToElectArmy({ you, opponent }) {
+    return this.withMutations(mnt => {
+      mnt.set('state', STATE.get('PREPARATION'))
+        .set('player', you)
+        .set('opponent', opponent);
+    });
+  }
+
+  startToLineupArmy(data) {
+    return this.withMutations(mnt => {
+      mnt.set('state', STATE.get('GAME'))
+        .set('controller', new Controller().set('offense', this.player.offense).sync(data));
+    });
+  }
+
+  syncData(data) {
+    return this.set('controller', this.controller.sync(data));
+  }
+
+  rejectAction() {
+    return this.set('controller', this.controller.clearUI());
+  }
+
+  selectCell(cellId, socket) {
+    return this.set('controller', 
+      this.controller.selectCell(cellId, (from, to, target) => {
+        socket.emit('act', {
+          from: from,
+          to: to,
+          target: target
+        });
+      })
+    );
+  }
+
+  hoverCell(cellId) {
+    return this.set('controller', this.controller.hoverCell(cellId));
   }
 
 }
