@@ -2,13 +2,10 @@ const express = require('express');
 const Path = require('path');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
-const session = require('express-session');
-const RedisStore = require('connect-redis')(session);
-const redisStore = new RedisStore({});
 const app = express();
-const SECRET = 'your secret';
 const uid = require('uid-safe').sync;
-const redis = require('./redis.js');
+const jwt = require('jsonwebtoken');
+const JWT_SECRET = 'qjkrfelqjfkaJ';
 
 app.set('view engine', 'pug');
 app.set('views', Path.join(__dirname, '../views'));
@@ -16,40 +13,38 @@ app.use(express.static(Path.join(__dirname, '../public')));
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 app.use(cookieParser());
-app.use(session({
-  store: redisStore,
-  secret: SECRET,
-  resave: false,
-  saveUninitialized: true,
-  cookie: {
-  }
-}));
 
-const MatchServer = require('./game/server/MatchServer.js');
-const matchServer = new MatchServer();
+const GameServer = require('./game/server/Server.js');
+const gameServer = new GameServer();
 
 app.get('/', (req, res) => {
   res.render('index', {
     units: require('./game/data/json/unit.json')
   });
 });
-app.post('/matching', (req, res) => {
-  req.session.reload(() => {
-    req.session.userId = uid(24);
-    req.session.deck = req.body.deck;
-    res.render('matching');
-  });
+app.post('/app', (req, res) => {
+  res.cookie('jwt', jwt.sign({
+    userId: uid(24),
+    deck: Array.prototype.concat([], req.body.deck)
+  }, JWT_SECRET));
+  res.render('app');
+});
+app.get('/test', (req, res) => {
+  res.render('test');
 });
 
 app.get('/game/:id', (req, res, next) => {
   const mid = req.params.id;
-  if (!matchServer.existsMatch(mid)) {
+  if (!gameServer.existsMatch(mid)) {
     return next(new Error('game not found'));
   }
   res.render('game', {mid: mid});
 });
 app.get('/game/:id/stat', (req, res) => {
   res.send('data');
+});
+app.get('/favicon.ico', function(req, res) {
+  res.send(204);
 });
 
 app.use((req, res, next) => {
@@ -65,90 +60,18 @@ const server = require('http').createServer(app);
 server.listen(3005);
 
 const io = require('socket.io').listen(server);
-const handshake = function(socket, next) {
-  const cookie = require('cookie').parse(socket.request.headers.cookie);
-  const sid = require('cookie-parser').signedCookie(cookie['connect.sid'], SECRET);
-  if (sid) {
-    redisStore.get(sid, (err, sess) => {
-      if (!err && sess && sess.userId) {
-        socket.userId = sess.userId;
-        socket.deck = sess.deck;
-        return next();
-      } else {
-        next('fail', false);
-      }
-    });
-  } else {
-    next('fail', false);
-  }
-};
-
-const matchmaker = require('./matchmaker.js');
-
-const matchingNS = io.of('/matching');
-matchingNS.use(handshake);
-matchingNS.on('connection', socket => {
-  matchmaker.wait(socket.id);
-  socket.on('disconnect', () => {
-    matchmaker.remove(socket.id);
-  });
-});
-
-const bredis = redis.duplicate();
-function makeMatch() {
-  bredis.brpop('matching', 0, (err, replies) => {
-    const ids = JSON.parse(replies[1]);
-    //TODO heart beat & retry
-    const match = matchServer.createMatch();
-    ids.forEach(socketId => {
-      const socket = matchingNS.connected[socketId];
-      match.player.add(socket.userId, socket.deck);
-      socket.emit('done', match.id);
-    });
-
-    makeMatch();
-  });
-}
-makeMatch();
-
-
 const gameNS = io.of('/game');
-gameNS.use(handshake);
-gameNS.on('connection', socket => {
-  socket.on('join', matchId => {
-    const match = matchServer.getMatch(matchId);
-    if (!match) {
-      return;
+gameNS.use((socket, next) => {
+  const cookie = require('cookie').parse(socket.request.headers.cookie);
+  jwt.verify(cookie.jwt, JWT_SECRET, (err, data) => {
+    if (err) {
+      return next('fail', false);
     }
-    socket.join(matchId);
-    const userId = socket.userId;
-    socket.emit('metaData', match.metaData(userId));
-    socket.emit('mirror', match.game.data());
-
-    if (!match.player.isPlayer(userId)) {
-      return;
-    }
-
-    if (!match.player.isReady(userId)) {
-      socket.emit('preparation');
-      socket.on('prepared', selectedIndexes => {
-        match.player.setSortie(userId, selectedIndexes);
-        if (match.engage()) {
-          gameNS.to(matchId).emit('mirror', match.game.data());
-        }
-      });
-    }
-
-    socket.on('action', (fromCid, toCid, targetCid) => {
-      if (match.action(userId, fromCid, toCid, targetCid)) {
-        gameNS.to(matchId).emit('completeAction', match.game.data());
-        const winnedPnum = match.winnedPnum(match.game);
-        if (winnedPnum !== undefined) {
-          gameNS.to(matchId).emit('winner', winnedPnum);
-        }
-      }
-    });
-
+    console.log('user', data.userId);
+    socket.userId = data.userId;
+    socket.deck = data.deck;
+    return next();
   });
 });
+gameServer.initSocket(gameNS);
 
