@@ -1,14 +1,11 @@
 
 const path = require('path');
-// var webpack = require('webpack');
 
-const ExtractTextPlugin = require('extract-text-webpack-plugin'); 
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 module.exports = {
   entry: {
-    game: './src/game/client/Container.jsx',
-    matching: './src/public/matching/index.js',
-    // game2: './src/game/client/Container.jsx'
+    app: './src/game/client/Container.jsx',
   },
   output: {
     path: path.resolve('public'),
@@ -19,19 +16,23 @@ module.exports = {
     rules: [
       {
         test: /\.(js|jsx)$/,
-        loader: 'babel-loader'
+        use: ['babel-loader']
       },
       {
         test: /\.css$/,
         use: ExtractTextPlugin.extract({
           use: 'css-loader'
         })
-      }
+      },
+      {
+        test: /\.(jpg|png)$/,
+        use: ['file-loader']
+      },
     ]
   },
   plugins: [
-    new ExtractTextPlugin('bundle.css')
-  ]
+    new ExtractTextPlugin('bundle.css'),
+  ],
   // plugins: [
     // new webpack.optimize.UglifyJsPlugin({
       // compress: {
@@ -39,4 +40,8 @@ module.exports = {
       // }
     // })
   // ]
+  devServer: {
+    port: 8081,
+    contentBase: path.join(__dirname, 'public')
+  }
 };
